Cleanup underwater checks, fix rectangular water bubbles
Long/thin rectangular nodes were not handled correctly,
because it would only check the nearest waterbubble node
for overlap, not all on the map.

+ Lua func: isUnderWater(x, y, radius)

+++ BBGE/Rect.h
@@ -59,7 +59,7 @@ public:
 		*x = x1 + ((*w) / 2);
 		*y = y1 + ((*h) / 2);
 	}
-	bool isCoordinateInside(const Vector &vec, int radius=0) const
+	bool isCoordinateInside(const Vector &vec, float radius=0) const
 	{
 		return ((vec.x >= x1-radius && vec.x <= x2+radius) && (vec.y >= y1-radius && vec.y <= y2+radius));
 	}
